move functions from common data file to forecast data file and update path imports
